refactor(upload): centralize extension normalization helpers
